Cheats and hacks
There was a bug somwhere in the code, whereby Dragon Tap blocks-as-items
were showing up as the purple-grey-block-of-sadness. So, I made a new
block, had it mimic the behavior of the standard item block, and set it
to drop from the Dragon Tap statues.

There is no way that this will come back to bite me. Nope, not any way
at all.