Misc Improvements
The dragon wand requires ten experience levels before it will allow you
to set a dragon tap (and then takes those ten levels from you as
payment). Also, new texture for it. If you're digging with a stone
trowel and a rune would drop, nothing else will along with it.